pyproject.toml
[project]
name = "whodini"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "annotated-doc==0.0.4",
    "annotated-types==0.7.0",
    "anyio==4.12.1",
    "certifi==2026.1.4",
    "charset-normalizer==3.4.4",
    "click==8.3.1",
    "colorama==0.4.6",
    "contourpy==1.3.3",
    "cycler==0.12.1",
    "fastapi==0.135.1",
    "filelock==3.20.3",
    "fonttools==4.61.1",
    "fsspec==2026.1.0",
    "h11==0.16.0",
    "httptools==0.7.1",
    "idna==3.11",
    "insightface>=0.7.3",
    "ipykernel>=7.2.0",
    "jinja2==3.1.6",
    "kiwisolver==1.4.9",
    "lap==0.5.13",
    "markupsafe==3.0.3",
    "matplotlib==3.10.8",
    "mpmath==1.3.0",
    "networkx==3.6.1",
    "numpy==2.3.5",
    "onnxruntime>=1.24.3",
    "opencv-python==4.13.0.90",
    "packaging==26.0",
    "pandas==2.3.3",
    "pillow==12.1.0",
    "polars==1.37.1",
    "polars-runtime-32==1.37.1",
    "psutil==7.2.2",
    "pydantic==2.12.5",
    "pydantic-core==2.41.5",
    "pyparsing==3.3.2",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.2.2",
    "python-multipart==0.0.22",
    "pytz==2025.2",
    "pyyaml==6.0.3",
    "requests==2.32.5",
    "scipy==1.17.0",
    "setuptools==80.10.2",
    "six==1.17.0",
    "starlette==0.52.1",
    "sympy==1.14.0",
    "torch==2.9.1",
    "torchvision==0.24.1",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.2",
    "tzdata==2025.2",
    "ultralytics==8.4.8",
    "ultralytics-thop==2.0.18",
    "urllib3==2.6.3",
    "uvicorn==0.42.0",
    "watchfiles==1.1.1",
    "websockets==16.0",
]
